updates to ci setup, detecting if miniconda is already installed
(trying to speed up builds esp on buildkite agents)

+++ ci/travis-setup.sh
@@ -2,10 +2,15 @@
 set -eo pipefail
 set -x
 
-# Sets up travis-ci environment for testing bioconda-utils.
-curl -O https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
-bash Miniconda3-latest-Linux-x86_64.sh -f -b -p $HOME/anaconda
-export PATH=$HOME/anaconda/bin:$PATH
+CONDA_DIR=$HOME/anaconda
+ENVNAME=lcdb-wf-test
+
+# Install miniconda if it doesn't already exist.
+if [[ -e $CONDA_DIR ]]; then
+    curl -O https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh
+    bash Miniconda3-latest-Linux-x86_64.sh -f -b -p $CONDA_DIR
+    export PATH="$CONDA_DIR/bin:$PATH"
+fi
 
 # Add channels in the specified order.
 conda config --add channels conda-forge
@@ -14,9 +19,18 @@ conda config --add channels r
 conda config --add channels bioconda
 conda config --add channels lcdb
 
-ENVNAME=lcdb-wf-test
-conda env list | grep -q $ENVNAME && conda env remove -y -n $ENVNAME
-conda create -n $ENVNAME -y python=3.5 --file requirements.txt | tee setup.log | grep -v " Time: "
-source activate $ENVNAME
+# Environment exists; update with requirements. This can happen on a buildkite
+# agent.
+if [[ conda env list | grep -q $ENVNAME ]]; then
+    conda install -y --file requirements.txt python=3.5
+else
+    # Otherwise create a new environment (this will happen every time on
+    # travis-ci). Don't show the progress for package downloads in the main
+    # log, but provide setup.log as a build artifact for buildkite.
+    conda create -n $ENVNAME -y python=3.5 --file requirements.txt \
+        | tee setup.log \
+        | grep -v " Time: "
+fi
 
+source activate $ENVNAME
 python ci/get-data.py
